# Project encoding performance improvements