--- a/Sources/xcodeproj/Extensions/Dictionary+Enumerate.swift
+++ b/Sources/xcodeproj/Extensions/Dictionary+Enumerate.swift
@@ -7,7 +7,10 @@ extension Dictionary {
         using block: (Any, Any, UnsafeMutablePointer<ObjCBool>) throws -> Void) throws
     {
         var blockError: Error?
-        (self as NSDictionary).enumerateKeysAndObjects(options: opts) { (key, obj, stops) in
+        // For performance it is very important to create a separate dictionary instance.
+        // (self as NSDictionary).enumerateKeys... - works much slower
+        let dictionary = NSDictionary(dictionary: self)
+        dictionary.enumerateKeysAndObjects(options: opts) { (key, obj, stops) in
             do {
                 try block(key, obj, stops)
             } catch {