--- a/Sources/xcodeproj/Objects/BuildPhase/PBXBuildFile.swift
+++ b/Sources/xcodeproj/Objects/BuildPhase/PBXBuildFile.swift
@@ -108,22 +108,36 @@ extension PBXBuildFile {
 }
 
 // MARK: - PlistSerializable
+// Helper for serialize the BuildFile with associated BuildPhase
+final class PBXBuildPhaseFile: PlistSerializable, Equatable {
 
-extension PBXBuildFile: PlistSerializable {
     var multiline: Bool { return false }
-
+    
+    let buildFile: PBXBuildFile
+    let buildPhase: PBXBuildPhase
+    
+    init(buildFile: PBXBuildFile, buildPhase: PBXBuildPhase) {
+        self.buildFile = buildFile
+        self.buildPhase = buildPhase
+    }
+    
     func plistKeyAndValue(proj _: PBXProj, reference: String) throws -> (key: CommentedString, value: PlistValue) {
         var dictionary: [CommentedString: PlistValue] = [:]
         dictionary["isa"] = .string(CommentedString(PBXBuildFile.isa))
-        if let fileReference = fileReference {
+        if let fileReference = buildFile.fileReference {
             let fileElement: PBXFileElement? = fileReference.getObject()
             dictionary["fileRef"] = .string(CommentedString(fileReference.value, comment: fileElement?.fileName()))
         }
-        if let settings = settings {
+        if let settings = buildFile.settings {
             dictionary["settings"] = settings.plist()
         }
-        let comment = try buildPhaseName().flatMap({ "\(try fileName() ?? "(null)") in \($0)" })
+        let comment = try buildPhase.name().flatMap({ "\(try buildFile.fileName() ?? "(null)") in \($0)" })
         return (key: CommentedString(reference, comment: comment),
                 value: .dictionary(dictionary))
     }
+
+    static func == (lhs: PBXBuildPhaseFile, rhs: PBXBuildPhaseFile) -> Bool {
+        return lhs.buildFile == rhs.buildFile && lhs.buildPhase == rhs.buildPhase
+    }
+
 }