--- a/Tests/xcodeprojTests/Objects/Project/PBXOutputSettingsTests.swift
+++ b/Tests/xcodeprojTests/Objects/Project/PBXOutputSettingsTests.swift
@@ -9,6 +9,9 @@ class PBXOutputSettingsTsts: XCTestCase {
 
     var objectBuildFileAssets: (PBXObjectReference, PBXBuildFile)!
     var objectBuildFileMain: (PBXObjectReference, PBXBuildFile)!
+    
+    var objectBuildPhaseFileAssets: (PBXObjectReference, PBXBuildPhaseFile)!
+    var objectBuildPhaseFileMain: (PBXObjectReference, PBXBuildPhaseFile)!
 
     var fileReferenceAssets: PBXFileReference!
     var fileReferenceCoreData: PBXFileReference!
@@ -38,6 +41,9 @@ class PBXOutputSettingsTsts: XCTestCase {
 
         objectBuildFileAssets = (buildFileAssets.reference, buildFileAssets)
         objectBuildFileMain = (buildFileMain.reference, buildFileMain)
+        
+        objectBuildPhaseFileAssets = proj.objects.buildPhaseFile.first { $0.value.buildFile.file?.fileName() == "Assets.xcassets" }!
+        objectBuildPhaseFileMain = proj.objects.buildPhaseFile.first { $0.value.buildFile.file?.fileName() == "Main.storyboard" }!
 
         fileReferenceAssets = proj.fileReferences.first { $0.fileName() == "Assets.xcassets" }!
         fileReferenceCoreData = proj.fileReferences.first { $0.fileName() == "CoreData.framework" }!
@@ -82,6 +88,18 @@ class PBXOutputSettingsTsts: XCTestCase {
         buildFileMain.file = nil
         XCTAssertFalse(PBXFileOrder.byFilename.sort(lhs: (ref1, buildFileAssets), rhs: (ref2, buildFileMain)))
         XCTAssertTrue(PBXFileOrder.byFilename.sort(lhs: (ref2, buildFileMain), rhs: (ref1, buildFileAssets)))
+    }
+    
+    // MARK: - PBXFileOrder - PBXBuildPhaseFile
+    
+    func test_PBXFileOrder_PBXBuildPhaseFile_by_uuid() {
+        XCTAssertFalse(PBXFileOrder.byUUID.sort(lhs: objectBuildPhaseFileAssets, rhs: objectBuildPhaseFileMain))
+        XCTAssertTrue(PBXFileOrder.byUUID.sort(lhs: objectBuildPhaseFileMain, rhs: objectBuildPhaseFileAssets))
+    }
+    
+    func test_PBXFileOrder_PBXBuildPhaseFile_by_filename() {
+        XCTAssertTrue(PBXFileOrder.byFilename.sort(lhs: objectBuildPhaseFileAssets, rhs: objectBuildPhaseFileMain))
+        XCTAssertFalse(PBXFileOrder.byFilename.sort(lhs: objectBuildPhaseFileMain, rhs: objectBuildPhaseFileAssets))
     }
 
     // MARK: - PBXFileOrder - PBXFileReference