## Improve project encoding performance 📝
For my test project: `xcodeproject.write(path: path)` was 57 seconds, and it became 8 seconds.

## Improvements
## Obtain PBXBuildFile from PBXBuildPhase instead searching PBXBuildPhase in PBXObjects
The main blocker was obtaining build phase name for PBXBuildFile encoding from PBXObjects -`plistKeyAndValue`. For each file, we ran through all the build phases and searched for a file with the same identifier. Of course, it took a lot of time.

Instead, I added an utility class that contains the file and build phase from which this file is. I am building this class from the build phase files.

This speed up for encoding on 46 seconds. ( 11 seconds instead of 57 ).

## Replaced the append in the string to append into the array + join.
Append into a typed array faster than into a string. This gave about 3 seconds. ( 8 seconds instead of 11 ).

Also this preparation for the subsequent changes. For an array, we can pre-set the size. This will allow non-allocation of memory several times. And we can prepare the data for recording in parallel.

Now this code slows down the most:
```
        elements.forEach { (element, multiline) in
            write(dictionaryKey: element.key, dictionaryValue: element.value, multiline: multiline)
        }
```
We can pass an inout array for each thread. I'm working on it now.