@CognitiveDisson 2019-02-20
@pepibumur I fixed all issues. And added one more optimization - concurrent decoding. Performance boost from 8 second to ~3.